backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
openai>=1.12.0
nacos-sdk-python>=0.1.13
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-grpc>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
grpcio>=1.60.0
grpcio-tools>=1.60.0
aiohttp>=3.9.3
requests>=2.31.0
chromadb>=0.4.22
pymysql>=1.1.0
aiomysql>=0.2.0
redis>=5.0.1
celery>=5.3.6
flower>=2.0.1
loguru>=0.7.2
tenacity>=8.2.3
sqlalchemy>=2.0.25
alembic>=1.13.1
pika>=1.3.2
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pypdf>=4.0.0
